refactor tests, and simplify n5 to_dask at expense of some DRY sins
