+++ src/fibsem_tools/io/n5/__init__.py
@@ -10,11 +10,47 @@ from xarray import DataArray
 from zarr.errors import ReadOnlyError
 
 from fibsem_tools.types import PathLike
+from fibsem_tools.io.n5.hierarchy import cosem, neuroglancer
+from ..zarr import access as access_zarr, get_url
+import dask.array as da
+from dask.base import tokenize
 
-from ..zarr import access as access_zarr
-from ..zarr import to_dask
 
-N5_AXES_3D = ["x", "y", "z"]
+def to_dask(
+    arr: zarr.Array,
+    chunks: Literal["auto"] | Sequence[int] = "auto",
+    inline_array: bool = True,
+    **kwargs: Any,
+) -> da.Array:
+    """
+    Create a Dask array from a Zarr array. This is a very thin wrapper around `dask.array.from_array`
+
+    Parameters
+    ----------
+
+    arr : zarr.Array
+
+    chunks : Literal['auto'] | Sequence[int]
+        The chunks to use for the output Dask array. It may be tempting to use the chunk size of the
+        input array for this parameter, but be advised that Dask performance suffers when arrays
+        have too many chunks, and Zarr arrays routinely have too many chunks by this definition.
+
+    inline_array : bool, default is `True`
+        Whether the Zarr array should be inlined in the Dask compute graph. See documentation for
+        `dask.array.from_array` for details.
+
+    **kwargs : Any
+        Additional keyword arguments for `dask.array.from_array`
+
+    Returns
+    -------
+
+    dask.array.Array
+    """
+    if kwargs.get("name") is None:
+        kwargs["name"] = f"{get_url(arr)}-{tokenize(arr)}"
+    darr = da.from_array(arr, chunks=chunks, inline_array=inline_array, **kwargs)
+    return darr
 
 
 class N5FSStorePatched(zarr.N5FSStore):
@@ -44,9 +80,6 @@ DEFAULT_N5_STORE = N5FSStorePatched
 def access(store: PathLike, path: PathLike, **kwargs: Any) -> zarr.Group | zarr.Array:
     store = DEFAULT_N5_STORE(store, **kwargs.get("storage_options", {}))
     return access_zarr(store, path, **kwargs)
-
-
-from fibsem_tools.io.n5.hierarchy import cosem, neuroglancer
 
 
 def create_dataarray(
