+++ src/fibsem_tools/io/n5/hierarchy/neuroglancer.py
@@ -15,8 +15,9 @@ from xarray import DataArray
 
 from fibsem_tools.chunk import normalize_chunks
 from fibsem_tools.coordinate import stt_coord, stt_from_array
-from fibsem_tools.io.n5 import N5_AXES_3D
 from fibsem_tools.io.zarr import access_parent
+
+N5_AXES_3D = ["x", "y", "z"]
 
 
 def model_group(
