+++ tests/io/test_dask.py
@@ -5,7 +5,7 @@ from typing import TYPE_CHECKING
 from fibsem_tools.io.core import read
 
 if TYPE_CHECKING:
-    from typing import Tuple, Union
+    from typing import Tuple
 
 import dask
 import dask.array as da
@@ -14,10 +14,6 @@ import pytest
 import zarr
 from dask.array.core import slices_from_chunks
 from fibsem_tools.chunk import (
-    autoscale_chunk_shape,
-    ensure_minimum_chunksize,
-    interval_remainder,
-    resolve_slice,
     resolve_slices,
 )
 from fibsem_tools.io.dask import (
@@ -28,55 +24,6 @@ from fibsem_tools.io.dask import (
 )
 from numpy.testing import assert_array_equal
 from pydantic_zarr.v2 import ArraySpec, GroupSpec
-
-
-def test_ensure_minimum_chunksize():
-    data = da.zeros((10,), chunks=(2,))
-    assert ensure_minimum_chunksize(data, (4,)).chunksize == (4,)
-
-    data = da.zeros((10,), chunks=(6,))
-    assert ensure_minimum_chunksize(data, (4,)).chunksize == (6,)
-
-    data = da.zeros((10, 10), chunks=(2, 1))
-    assert ensure_minimum_chunksize(data, (4, 4)).chunksize == (4, 4)
-
-    data = da.zeros((10, 10, 10), chunks=(2, 1, 10))
-    assert ensure_minimum_chunksize(data, (4, 4, 4)).chunksize == (4, 4, 10)
-
-
-def test_autoscale_chunk_shape():
-    chunk_shape = (1,)
-    array_shape = (1000,)
-    size_limit = "1KB"
-    dtype = "uint8"
-
-    assert autoscale_chunk_shape(chunk_shape, array_shape, size_limit, dtype) == (999,)
-
-    chunk_shape = (1, 1)
-    array_shape = (1000, 100)
-
-    assert autoscale_chunk_shape(chunk_shape, array_shape, size_limit, dtype) == (
-        999,
-        1,
-    )
-
-    chunk_shape = (1, 1)
-    array_shape = (1000, 1001)
-
-    assert autoscale_chunk_shape(chunk_shape, array_shape, size_limit, dtype) == (
-        1,
-        999,
-    )
-
-    chunk_shape = (64, 64, 64)
-    size_limit = "8MB"
-    array_shape = (1000, 1000, 1000)
-
-    assert autoscale_chunk_shape(chunk_shape, array_shape, size_limit, dtype) == (
-        64,
-        64,
-        1024,
-    )
 
 
 @pytest.mark.parametrize("keep_attrs", (True, False))
@@ -125,44 +72,6 @@ def test_write_blocks_delayed():
     result = dask.compute(w_ops)[0]
     assert result == slices_from_chunks(arr.chunks)
     assert_array_equal(np.array(arr), z_arr)
-
-
-@pytest.mark.parametrize(
-    "data, expected",
-    [
-        (((0, 1), (0, 1)), (0, 0)),
-        (((0, 2), (0, 1)), (0, 0)),
-        (((1, 2), (0, 1)), (1, 0)),
-        (((0, 10), (0, 1)), (0, 0)),
-    ],
-)
-def test_interval_remainder(data, expected):
-    assert interval_remainder(*data) == expected
-
-
-@pytest.mark.parametrize(
-    "data, expected",
-    [
-        (
-            (slice(None), (0, 10)),
-            slice(0, 10, 1),
-        ),
-        (
-            (slice(0, 10, 1), (0, 10)),
-            slice(0, 10, 1),
-        ),
-        (
-            (slice(0, 10), (0, 10)),
-            slice(0, 10, 1),
-        ),
-        (
-            (slice(None), (9, 10)),
-            slice(9, 10, 1),
-        ),
-    ],
-)
-def test_resolve_slice(data, expected):
-    assert resolve_slice(*data) == expected
 
 
 @pytest.mark.parametrize(
